Improve performance and use 'case' instead of 'if's
Make initialization function run in the background.

As it had to ping https://archlinux.org to check if it was reachable, it
made the program noticeably lag a bit when it was invoked.

Use 'case' statements instead of 'if' statements, as it makes code for
handling input easier to comprehend.